Remove the games as a separate package and instead make it a relative import
