+++ packages/frontend/package.json
@@ -14,7 +14,6 @@
     "@radix-ui/themes": "^3.1.3",
     "@reduxjs/toolkit": "^2.2.7",
     "@resync-games/api": "workspace:^",
-    "@resync-games/games": "workspace:^",
     "@resync-games/redux-store": "workspace:^",
     "clsx": "^2.1.1",
     "lodash-es": "^4.17.21",
