+++ packages/frontend/src/app/games/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateGame from "@/components/createGame/createGame";
+import CreateGame from "@/components/createGame/CreateGame";
 
 export default function NavigateToGameCreate() {
   return <CreateGame />;
