--- packages/frontend/src/lib/game/baseScene.ts
@@ -1,7 +1,0 @@
-import { Scene as PhaserScene } from "phaser";
-
-export class BaseScene extends PhaserScene {
-  constructor(private sceneName: string) {
-    super();
-  }
-}
