+++ packages/frontend/tsconfig.json
@@ -27,9 +27,11 @@
     "jsx": "preserve",
     "incremental": true,
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@resync-games/games/*": ["../games/src/frontend/*"],
+      "@resync-games/games-shared/*": ["../games/src/shared/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "../games/src/frontend/**/*.tsx", "../games/src/shared/**/*.ts", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
 }
