+++ packages/games/src/frontend/frontendRegistry.ts
@@ -1,5 +1,5 @@
 import { GameInfo } from "@resync-games/api";
-import { GAME_SLUGS } from "../gamesRegistry";
+import { GAME_SLUGS } from "../shared/gamesRegistry";
 import { PongHomePage } from "./pong/pong";
 import { DisplayMagicMazeGame } from "./snatchTheSnack/snatchTheSnack";
 import { IGameStateHandler } from "@resync-games/redux-store";
