+++ packages/games/src/frontend/snatchTheSnack/components/PawnMovement.module.scss
