+++ packages/games/src/frontend/snatchTheSnack/components/PawnMovement.tsx
