+++ packages/games/src/frontend/snatchTheSnack/components/SelectPawn.module.scss
