+++ packages/games/src/frontend/snatchTheSnack/components/SelectPawn.tsx
