+++ packages/games/src/shared/gamesRegistry.ts
